Exercice5 fini

+++ House.cs
@@ -1,7 +1,7 @@
 public class House
 {
-    public static double stone_needed =3;
-    public static double wood_needed =3;
+    public static int stone_needed =3;
+    public static int wood_needed =3;
     public static int villageois =10;
